big reorg, now a single source of install scripts, updated scripts, updated Vagrantfile with some networking config that sometimes helps

+++ install/install-00-base.sh
@@ -13,8 +13,15 @@ sudo apt-get install build-essential -y
 # sudo apt-get install lxde -y
 
 # install zsh
-git clone git://github.com/robbyrussell/oh-my-zsh.git /home/vagrant/.oh-my-zsh
+git clone https://github.com/robbyrussell/oh-my-zsh.git /home/vagrant/.oh-my-zsh
 cp /home/vagrant/.oh-my-zsh/templates/zshrc.zsh-template /home/vagrant/.zshrc
 sudo chsh -s /bin/zsh vagrant
 
-sudo cp /vagrant/resources/hosts /etc
+if [ -f /vagrant/resources/hosts ]; then
+    sudo cp /vagrant/resources/hosts /etc
+fi
+
+if [ -d /vagrant/resources/.ssh ]; then
+    su vagrant
+    cp -r /vagrant/resources/.ssh /home/vagrant/.ssh
+fi
